requirements.txt
APScheduler==3.10.4
asgiref==3.7.2
colorama==0.4.6
colorlog==6.7.0
Django==3.2.18
django-apscheduler==0.6.2
django-environ==0.10.0
django-extensions==3.2.1
django-saml2-pro-auth==1.0.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
isodate==0.6.1
lxml==4.9.2
mysqlclient==2.1.1
PyJWT==2.8.0
python3-saml==1.15.0
pytz==2023.3
six==1.16.0
sqlparse==0.4.4
typing_extensions==4.6.3
tzdata==2023.3
tzlocal==5.1
xmlsec==1.3.13